+++ server/src/lab9.rs
@@ -615,7 +615,7 @@ impl Labyrinth9 {
         let t = BANKS[bank_index as usize].temp;
 
         let mut correct = true;
-        let mut m = x + y * core::constants::SERVER_MAPX as i32;
+        let mut m = x + y * core::constants::SERVER_MAPX;
 
         for n in 0..core::constants::SWITCHES {
             let item_number = Repository::with_map(|map| map[m as usize].it);
@@ -646,7 +646,7 @@ impl Labyrinth9 {
 
         // Door logic
         m = self.banks[bank_index as usize].doorx
-            + self.banks[bank_index as usize].doory * core::constants::SERVER_MAPX as i32;
+            + self.banks[bank_index as usize].doory * core::constants::SERVER_MAPX;
 
         let item_number = Repository::with_map(|map| map[m as usize].it);
 
@@ -935,8 +935,8 @@ impl Labyrinth9 {
                 state.do_area_notify(
                     cn as i32,
                     0,
-                    character_position.0 as i32,
-                    character_position.1 as i32,
+                    character_position.0,
+                    character_position.1,
                     core::constants::NT_SEE as i32,
                     cn as i32,
                     0,
